Move api keys loading to separate hook
